lib,src: split up big builtin scripts
Reduces the body of parsed source code from 210 kb to 80 kb for a null
program (`node -e 0`) and from 255 kb to 190 kb for a one-liner that
prints some output to the console.

It doesn't seem to have much impact on the size of the generated code
(on my system I see an 8% reduction with --nolazy and 3% without) but
hopefully it speeds up startup on slow systems because there is now
less source code to parse.
