+++ lib/_stream_duplex.js
@@ -7,11 +7,11 @@
 
 module.exports = Duplex;
 
-const util = require('util');
 const Readable = require('_stream_readable');
 const Writable = require('_stream_writable');
+const inherits = require('internal/inherits');
 
-util.inherits(Duplex, Readable);
+inherits(Duplex, Readable);
 
 var keys = Object.keys(Writable.prototype);
 for (var v = 0; v < keys.length; v++) {
