+++ lib/_stream_passthrough.js
@@ -7,8 +7,8 @@
 module.exports = PassThrough;
 
 const Transform = require('_stream_transform');
-const util = require('util');
-util.inherits(PassThrough, Transform);
+const inherits = require('internal/inherits');
+inherits(PassThrough, Transform);
 
 function PassThrough(options) {
   if (!(this instanceof PassThrough))
