+++ lib/_stream_readable.js
@@ -3,14 +3,16 @@
 module.exports = Readable;
 Readable.ReadableState = ReadableState;
 
-const EE = require('events').EventEmitter;
-const Stream = require('stream');
 const Buffer = require('buffer').Buffer;
-const util = require('util');
-const debug = util.debuglog('stream');
+const EE = require('events').EventEmitter;
+const Stream = require('internal/stream');
+const debug = require('internal/debuglog')('stream');
+const inherits = require('internal/inherits');
+
+var Duplex;  // Break recursive dependency with _stream_duplex.js.
 var StringDecoder;
 
-util.inherits(Readable, Stream);
+inherits(Readable, Stream);
 
 function ReadableState(options, stream) {
   options = options || {};
@@ -19,7 +21,10 @@ function ReadableState(options, stream) {
   // make all the buffer merging and length checks go away
   this.objectMode = !!options.objectMode;
 
-  if (stream instanceof Stream.Duplex)
+  if (Duplex === undefined)
+    Duplex = require('_stream_duplex');
+
+  if (stream instanceof Duplex)
     this.objectMode = this.objectMode || !!options.readableObjectMode;
 
   // the point at which it stops calling _read() to fill the buffer
