+++ lib/_stream_transform.js
@@ -45,8 +45,8 @@
 module.exports = Transform;
 
 const Duplex = require('_stream_duplex');
-const util = require('util');
-util.inherits(Transform, Duplex);
+const inherits = require('internal/inherits');
+inherits(Transform, Duplex);
 
 
 function TransformState(stream) {
