+++ lib/_stream_wrap.js
@@ -1,11 +1,11 @@
 'use strict';
 
 const assert = require('assert');
-const util = require('util');
 const Socket = require('net').Socket;
 const JSStream = process.binding('js_stream').JSStream;
 const uv = process.binding('uv');
-const debug = util.debuglog('stream_wrap');
+const debug = require('internal/debuglog')('stream_wrap');
+const inherits = require('internal/inherits');
 
 function StreamWrap(stream) {
   const handle = new JSStream();
@@ -57,7 +57,7 @@ function StreamWrap(stream) {
     handle: handle
   });
 }
-util.inherits(StreamWrap, Socket);
+inherits(StreamWrap, Socket);
 module.exports = StreamWrap;
 
 // require('_stream_wrap').StreamWrap
