+++ lib/_stream_writable.js
@@ -7,12 +7,14 @@
 module.exports = Writable;
 Writable.WritableState = WritableState;
 
-const util = require('util');
-const internalUtil = require('internal/util');
-const Stream = require('stream');
 const Buffer = require('buffer').Buffer;
+const Stream = require('internal/stream');
+const inherits = require('internal/inherits');
+const internalUtil = require('internal/util');
 
-util.inherits(Writable, Stream);
+inherits(Writable, Stream);
+
+var Duplex;  // Break recursive dependency with _stream_duplex.js.
 
 function nop() {}
 
@@ -30,7 +32,10 @@ function WritableState(options, stream) {
   // contains buffers or objects.
   this.objectMode = !!options.objectMode;
 
-  if (stream instanceof Stream.Duplex)
+  if (Duplex === undefined)
+    Duplex = require('_stream_duplex');
+
+  if (stream instanceof Duplex)
     this.objectMode = this.objectMode || !!options.writableObjectMode;
 
   // the point at which write() starts returning false
@@ -130,8 +135,13 @@ Object.defineProperty(WritableState.prototype, 'buffer', {
 function Writable(options) {
   // Writable ctor is applied to Duplexes, though they're not
   // instanceof Writable, they're instanceof Readable.
-  if (!(this instanceof Writable) && !(this instanceof Stream.Duplex))
-    return new Writable(options);
+  if (!(this instanceof Writable)) {
+    if (Duplex === undefined)
+      Duplex = require('_stream_duplex');
+
+    if (!(this instanceof Duplex))
+      return new Writable(options);
+  }
 
   this._writableState = new WritableState(options, this);
