+++ lib/console.js
@@ -1,6 +1,7 @@
 'use strict';
 
-const util = require('util');
+const format = require('internal/format');
+
 
 function Console(stdout, stderr) {
   if (!(this instanceof Console)) {
@@ -33,7 +34,7 @@ function Console(stdout, stderr) {
 }
 
 Console.prototype.log = function() {
-  this._stdout.write(util.format.apply(this, arguments) + '\n');
+  this._stdout.write(format.apply(this, arguments) + '\n');
 };
 
 
@@ -41,15 +42,22 @@ Console.prototype.info = Console.prototype.log;
 
 
 Console.prototype.warn = function() {
-  this._stderr.write(util.format.apply(this, arguments) + '\n');
+  this._stderr.write(format.apply(this, arguments) + '\n');
 };
 
 
 Console.prototype.error = Console.prototype.warn;
 
 
+var extend;
+var inspect;
+
 Console.prototype.dir = function(object, options) {
-  this._stdout.write(util.inspect(object, util._extend({
+  if (extend === undefined) {
+    extend = require('internal/extend');
+    inspect = require('util').inspect;
+  }
+  this._stdout.write(inspect(object, extend({
     customInspect: false
   }, options)) + '\n');
 };
@@ -75,7 +83,7 @@ Console.prototype.trace = function trace() {
   // exposed.
   var err = new Error();
   err.name = 'Trace';
-  err.message = util.format.apply(this, arguments);
+  err.message = format.apply(this, arguments);
   Error.captureStackTrace(err, trace);
   this.error(err.stack);
 };
@@ -84,7 +92,7 @@ Console.prototype.trace = function trace() {
 Console.prototype.assert = function(expression) {
   if (!expression) {
     var arr = Array.prototype.slice.call(arguments, 1);
-    require('assert').ok(false, util.format.apply(this, arr));
+    require('assert').ok(false, format.apply(this, arr));
   }
 };
