+++ lib/dgram.js
@@ -2,9 +2,12 @@
 
 const assert = require('assert');
 const Buffer = require('buffer').Buffer;
-const util = require('util');
 const events = require('events');
 const constants = require('constants');
+
+const errnoException = require('internal/errno_exception');
+const exceptionWithHostPort = require('internal/exception_with_host_port');
+const inherits = require('internal/inherits');
 
 const UDP = process.binding('udp_wrap').UDP;
 const SendWrap = process.binding('udp_wrap').SendWrap;
@@ -16,9 +19,6 @@ const BIND_STATE_BOUND = 2;
 // lazily loaded
 var cluster = null;
 var dns = null;
-
-const errnoException = util._errnoException;
-const exceptionWithHostPort = util._exceptionWithHostPort;
 
 function lookup(address, family, callback) {
   if (!dns)
@@ -101,7 +101,7 @@ function Socket(type, listener) {
   if (typeof listener === 'function')
     this.on('message', listener);
 }
-util.inherits(Socket, events.EventEmitter);
+inherits(Socket, events.EventEmitter);
 exports.Socket = Socket;
