+++ lib/fs.js
@@ -4,7 +4,6 @@
 'use strict';
 
 const SlowBuffer = require('buffer').SlowBuffer;
-const util = require('util');
 const pathModule = require('path');
 
 const binding = process.binding('fs');
@@ -34,7 +33,9 @@ const O_WRONLY = constants.O_WRONLY || 0;
 const isWindows = process.platform === 'win32';
 
 const DEBUG = process.env.NODE_DEBUG && /fs/.test(process.env.NODE_DEBUG);
-const errnoException = util._errnoException;
+const errnoException = require('internal/errno_exception');
+const extend = require('internal/extend');
+const inherits = require('internal/inherits');
 
 function throwOptionsError(options) {
   throw new TypeError('Expected options to be either an object or a string, ' +
@@ -1048,7 +1049,7 @@ function toUnixTimestamp(time) {
   if (typeof time === 'number') {
     return time;
   }
-  if (util.isDate(time)) {
+  if (Object.prototype.toString.call(time) === '[object Date]') {
     // convert to 123.456 UNIX timestamp
     return time.getTime() / 1000;
   }
@@ -1185,7 +1186,7 @@ fs.appendFile = function(path, data, options, callback_) {
   }
 
   if (!options.flag)
-    options = util._extend({ flag: 'a' }, options);
+    options = extend({ flag: 'a' }, options);
   fs.writeFile(path, data, options, callback);
 };
 
@@ -1198,7 +1199,7 @@ fs.appendFileSync = function(path, data, options) {
     throwOptionsError(options);
   }
   if (!options.flag)
-    options = util._extend({ flag: 'a' }, options);
+    options = extend({ flag: 'a' }, options);
 
   fs.writeFileSync(path, data, options);
 };
@@ -1219,7 +1220,7 @@ function FSWatcher() {
     }
   };
 }
-util.inherits(FSWatcher, EventEmitter);
+inherits(FSWatcher, EventEmitter);
 
 FSWatcher.prototype.start = function(filename, persistent, recursive) {
   nullCheck(filename);
@@ -1289,7 +1290,7 @@ function StatWatcher() {
     self.emit('stop');
   };
 }
-util.inherits(StatWatcher, EventEmitter);
+inherits(StatWatcher, EventEmitter);
 
 
 StatWatcher.prototype.start = function(filename, persistent, interval) {
@@ -1319,7 +1320,7 @@ fs.watchFile = function(filename, options, listener) {
   };
 
   if (options !== null && typeof options === 'object') {
-    options = util._extend(defaults, options);
+    options = extend(defaults, options);
   } else {
     listener = options;
     options = defaults;
@@ -1607,7 +1608,7 @@ fs.createReadStream = function(path, options) {
   return new ReadStream(path, options);
 };
 
-util.inherits(ReadStream, Readable);
+inherits(ReadStream, Readable);
 fs.ReadStream = ReadStream;
 
 function ReadStream(path, options) {
@@ -1779,7 +1780,7 @@ fs.createWriteStream = function(path, options) {
   return new WriteStream(path, options);
 };
 
-util.inherits(WriteStream, Writable);
+inherits(WriteStream, Writable);
 fs.WriteStream = WriteStream;
 function WriteStream(path, options) {
   if (!(this instanceof WriteStream))
@@ -1887,7 +1888,7 @@ function SyncWriteStream(fd, options) {
   this.autoClose = options.autoClose === undefined ? true : options.autoClose;
 }
 
-util.inherits(SyncWriteStream, Stream);
+inherits(SyncWriteStream, Stream);
 
 
 // Export
