+++ lib/internal/assert.js
@@ -1,0 +1,6 @@
+'use strict';
+
+module.exports = function assert(expression, message) {
+  if (!expression)
+    throw new Error(`Internal assertion failed: ${message || 'no message'}`);
+};
