+++ lib/internal/debuglog.js
@@ -1,0 +1,25 @@
+'use strict';
+
+const debugs = {};
+var debugEnviron;
+var format;
+
+module.exports = function(set) {
+  if (debugEnviron === undefined)
+    debugEnviron = process.env.NODE_DEBUG || '';
+  set = set.toUpperCase();
+  if (!debugs[set]) {
+    if (new RegExp('\\b' + set + '\\b', 'i').test(debugEnviron)) {
+      var pid = process.pid;
+      debugs[set] = function() {
+        if (format === undefined)
+          format = require('internal/format');
+        const msg = format.apply(exports, arguments);
+        console.error('%s %d: %s', set, pid, msg);
+      };
+    } else {
+      debugs[set] = function() {};
+    }
+  }
+  return debugs[set];
+};
