+++ lib/internal/errno_exception.js
@@ -1,0 +1,17 @@
+'use strict';
+
+var uv;
+
+module.exports = function errnoException(err, syscall, original) {
+  if (uv === undefined)
+    uv = process.binding('uv');
+  var errname = uv.errname(err);
+  var message = syscall + ' ' + errname;
+  if (original)
+    message += ' ' + original;
+  var e = new Error(message);
+  e.code = errname;
+  e.errno = errname;
+  e.syscall = syscall;
+  return e;
+};
