+++ lib/internal/extend.js
@@ -1,0 +1,13 @@
+'use strict';
+
+module.exports = function extend(origin, add) {
+  // Don't do anything if add isn't an object
+  if (add === null || typeof add !== 'object') return origin;
+
+  var keys = Object.keys(add);
+  var i = keys.length;
+  while (i--) {
+    origin[keys[i]] = add[keys[i]];
+  }
+  return origin;
+};
