+++ lib/internal/format.js
@@ -1,0 +1,50 @@
+'use strict';
+
+const formatRegExp = /%[sdj%]/g;
+module.exports = function format(f) {
+  if (typeof f !== 'string') {
+    var objects = [];
+    for (var i = 0; i < arguments.length; i++) {
+      objects.push(inspect(arguments[i]));
+    }
+    return objects.join(' ');
+  }
+
+  if (arguments.length === 1) return f;
+
+  var i = 1;
+  var args = arguments;
+  var len = args.length;
+  var str = String(f).replace(formatRegExp, function(x) {
+    if (x === '%%') return '%';
+    if (i >= len) return x;
+    switch (x) {
+      case '%s': return String(args[i++]);
+      case '%d': return Number(args[i++]);
+      case '%j':
+        try {
+          return JSON.stringify(args[i++]);
+        } catch (_) {
+          return '[Circular]';
+        }
+        // falls through
+      default:
+        return x;
+    }
+  });
+  for (var x = args[i]; i < len; x = args[++i]) {
+    if (x === null || (typeof x !== 'object' && typeof x !== 'symbol')) {
+      str += ' ' + x;
+    } else {
+      str += ' ' + inspect(x);
+    }
+  }
+  return str;
+};
+
+function inspect(x) {
+  if (inspect.cached === undefined)
+    inspect.cached = require('util').inspect;
+  return inspect.cached(x);
+}
+inspect.cached = undefined;
