+++ lib/module.js
@@ -1,14 +1,21 @@
 'use strict';
 
 const NativeModule = require('native_module');
-const util = require('util');
+const assert = require('internal/assert');
+const debug = require('internal/debuglog')('module');
 const internalUtil = require('internal/util');
 const runInThisContext = require('vm').runInThisContext;
-const assert = require('assert').ok;
-const fs = require('fs');
 const path = require('path');
 const internalModuleReadFile = process.binding('fs').internalModuleReadFile;
 const internalModuleStat = process.binding('fs').internalModuleStat;
+
+
+function fs() {
+  if (fs.cached === undefined)
+    fs.cached = require('fs');
+  return fs.cached;
+}
+fs.cached = undefined;
 
 
 // If obj.hasOwnProperty has been overridden, then calling
@@ -41,11 +48,7 @@ Module.globalPaths = [];
 
 Module.wrapper = NativeModule.wrapper;
 Module.wrap = NativeModule.wrap;
-Module._debug = util.debuglog('module');
-
-
-// We use this alias for the preprocessor that filters it out
-const debug = Module._debug;
+Module._debug = debug;
 
 
 // given a module name, and a list of paths to test, returns the first
@@ -106,7 +109,7 @@ function tryFile(requestPath) {
 }
 
 function toRealPath(requestPath) {
-  return fs.realpathSync(requestPath, Module._realpathCache);
+  return fs().realpathSync(requestPath, Module._realpathCache);
 }
 
 // given a path check a the file exists with any of the set extensions
@@ -448,14 +451,14 @@ function stripBOM(content) {
 
 // Native extension for .js
 Module._extensions['.js'] = function(module, filename) {
-  var content = fs.readFileSync(filename, 'utf8');
+  var content = fs().readFileSync(filename, 'utf8');
   module._compile(stripBOM(content), filename);
 };
 
 
 // Native extension for .json
 Module._extensions['.json'] = function(module, filename) {
-  var content = fs.readFileSync(filename, 'utf8');
+  var content = fs().readFileSync(filename, 'utf8');
   try {
     module.exports = JSON.parse(stripBOM(content));
   } catch (err) {
