+++ lib/net.js
@@ -1,13 +1,13 @@
 'use strict';
 
 const events = require('events');
-const stream = require('stream');
 const timers = require('timers');
-const util = require('util');
 const internalUtil = require('internal/util');
-const assert = require('assert');
 const cares = process.binding('cares_wrap');
 const uv = process.binding('uv');
+
+const Duplex = require('_stream_duplex');
+const Readable = require('_stream_readable');
 
 const Buffer = require('buffer').Buffer;
 const TTYWrap = process.binding('tty_wrap');
@@ -18,10 +18,13 @@ const PipeConnectWrap = process.binding('pipe_wrap').PipeConnectWrap;
 const ShutdownWrap = process.binding('stream_wrap').ShutdownWrap;
 const WriteWrap = process.binding('stream_wrap').WriteWrap;
 
+const assert = require('internal/assert');
+const debug = require('internal/debuglog')('net');
+const errnoException = require('internal/errno_exception');
+const exceptionWithHostPort = require('internal/exception_with_host_port');
+const inherits = require('internal/inherits');
 
 var cluster;
-const errnoException = util._errnoException;
-const exceptionWithHostPort = util._exceptionWithHostPort;
 
 function noop() {}
 
@@ -32,8 +35,6 @@ function createHandle(fd) {
   throw new TypeError('Unsupported fd type: ' + type);
 }
 
-
-const debug = util.debuglog('net');
 
 function isPipeName(s) {
   return typeof s === 'string' && toNumber(s) === false;
@@ -120,7 +121,7 @@ function Socket(options) {
   else if (options === undefined)
     options = {};
 
-  stream.Duplex.call(this, options);
+  Duplex.call(this, options);
 
   if (options.handle) {
     this._handle = options.handle; // private
@@ -170,7 +171,7 @@ function Socket(options) {
     }
   }
 }
-util.inherits(Socket, stream.Duplex);
+inherits(Socket, Duplex);
 
 Socket.prototype._unrefTimer = function unrefTimer() {
   for (var s = this; s !== null; s = s._parent)
@@ -280,9 +281,9 @@ exports.Stream = Socket; // Legacy naming.
 
 Socket.prototype.read = function(n) {
   if (n === 0)
-    return stream.Readable.prototype.read.call(this, n);
+    return Readable.prototype.read.call(this, n);
 
-  this.read = stream.Readable.prototype.read;
+  this.read = Readable.prototype.read;
   this._consuming = true;
   return this.read(n);
 };
@@ -397,7 +398,7 @@ Socket.prototype._read = function(n) {
 
 
 Socket.prototype.end = function(data, encoding) {
-  stream.Duplex.prototype.end.call(this, data, encoding);
+  Duplex.prototype.end.call(this, data, encoding);
   this.writable = false;
   DTRACE_NET_STREAM_END(this);
   LTTNG_NET_STREAM_END(this);
@@ -615,7 +616,7 @@ Socket.prototype.__defineGetter__('localPort', function() {
 Socket.prototype.write = function(chunk, encoding, cb) {
   if (typeof chunk !== 'string' && !(chunk instanceof Buffer))
     throw new TypeError('invalid data');
-  return stream.Duplex.prototype.write.apply(this, arguments);
+  return Duplex.prototype.write.apply(this, arguments);
 };
 
 
@@ -777,7 +778,7 @@ function connect(self, address, port, addressType, localAddress, localPort) {
   // TODO return promise from Socket.prototype.connect which
   // wraps _connectReq.
 
-  assert.ok(self._connecting);
+  assert(self._connecting);
 
   var err;
 
@@ -1033,7 +1034,7 @@ function afterConnect(status, handle, req, readable, writable) {
 
   debug('afterConnect');
 
-  assert.ok(self._connecting);
+  assert(self._connecting);
   self._connecting = false;
   self._sockname = null;
 
@@ -1113,7 +1114,7 @@ function Server(options, connectionListener) {
   this.allowHalfOpen = options.allowHalfOpen || false;
   this.pauseOnConnect = !!options.pauseOnConnect;
 }
-util.inherits(Server, events.EventEmitter);
+inherits(Server, events.EventEmitter);
 exports.Server = Server;
