+++ lib/path.js
@@ -1,12 +1,11 @@
 'use strict';
 
-const util = require('util');
 const isWindows = process.platform === 'win32';
 
 function assertPath(path) {
   if (typeof path !== 'string') {
     throw new TypeError('Path must be a string. Received ' +
-                        util.inspect(path));
+                        require('util').inspect(path));
   }
 }
