+++ lib/timers.js
@@ -2,9 +2,8 @@
 
 const Timer = process.binding('timer_wrap').Timer;
 const L = require('_linklist');
-const assert = require('assert').ok;
-const util = require('util');
-const debug = util.debuglog('timer');
+const assert = require('internal/assert');
+const debug = require('internal/debuglog')('timer');
 const kOnTimeout = Timer.kOnTimeout | 0;
 
 // Timeout values > TIMEOUT_MAX are set to 1.
