+++ lib/tty.js
@@ -1,12 +1,11 @@
 'use strict';
 
-const util = require('util');
 const internalUtil = require('internal/util');
 const net = require('net');
 const TTY = process.binding('tty_wrap').TTY;
 const isTTY = process.binding('tty_wrap').isTTY;
-const inherits = util.inherits;
-const errnoException = util._errnoException;
+const extend = require('internal/extend');
+const inherits = require('internal/inherits');
 
 
 exports.isatty = function(fd) {
@@ -28,7 +27,7 @@ function ReadStream(fd, options) {
   if (!(this instanceof ReadStream))
     return new ReadStream(fd, options);
 
-  options = util._extend({
+  options = extend({
     highWaterMark: 0,
     readable: true,
     writable: false,
@@ -79,6 +78,7 @@ WriteStream.prototype._refreshSize = function() {
   var winSize = [];
   var err = this._handle.getWindowSize(winSize);
   if (err) {
+    const errnoException = require('internal/errno_exception');
     this.emit('error', errnoException(err, 'getWindowSize'));
     return;
   }
