+++ lib/util.js
@@ -1,75 +1,12 @@
 'use strict';
 
-const uv = process.binding('uv');
 const Buffer = require('buffer').Buffer;
 const internalUtil = require('internal/util');
 var Debug;
 
-const formatRegExp = /%[sdj%]/g;
-exports.format = function(f) {
-  if (typeof f !== 'string') {
-    var objects = [];
-    for (var i = 0; i < arguments.length; i++) {
-      objects.push(inspect(arguments[i]));
-    }
-    return objects.join(' ');
-  }
-
-  if (arguments.length === 1) return f;
-
-  var i = 1;
-  var args = arguments;
-  var len = args.length;
-  var str = String(f).replace(formatRegExp, function(x) {
-    if (x === '%%') return '%';
-    if (i >= len) return x;
-    switch (x) {
-      case '%s': return String(args[i++]);
-      case '%d': return Number(args[i++]);
-      case '%j':
-        try {
-          return JSON.stringify(args[i++]);
-        } catch (_) {
-          return '[Circular]';
-        }
-        // falls through
-      default:
-        return x;
-    }
-  });
-  for (var x = args[i]; i < len; x = args[++i]) {
-    if (x === null || (typeof x !== 'object' && typeof x !== 'symbol')) {
-      str += ' ' + x;
-    } else {
-      str += ' ' + inspect(x);
-    }
-  }
-  return str;
-};
-
-
+exports.debuglog = require('internal/debuglog');
 exports.deprecate = internalUtil._deprecate;
-
-
-var debugs = {};
-var debugEnviron;
-exports.debuglog = function(set) {
-  if (debugEnviron === undefined)
-    debugEnviron = process.env.NODE_DEBUG || '';
-  set = set.toUpperCase();
-  if (!debugs[set]) {
-    if (new RegExp('\\b' + set + '\\b', 'i').test(debugEnviron)) {
-      var pid = process.pid;
-      debugs[set] = function() {
-        var msg = exports.format.apply(exports, arguments);
-        console.error('%s %d: %s', set, pid, msg);
-      };
-    } else {
-      debugs[set] = function() {};
-    }
-  }
-  return debugs[set];
-};
+exports.format = require('internal/format');
 
 
 /**
@@ -694,57 +631,9 @@ exports.log = function() {
 };
 
 
-/**
- * Inherit the prototype methods from one constructor into another.
- *
- * The Function.prototype.inherits from lang.js rewritten as a standalone
- * function (not on Function.prototype). NOTE: If this file is to be loaded
- * during bootstrapping this function needs to be rewritten using some native
- * functions as prototype setup using normal JavaScript does not work as
- * expected during bootstrapping (see mirror.js in r114903).
- *
- * @param {function} ctor Constructor function which needs to inherit the
- *     prototype.
- * @param {function} superCtor Constructor function to inherit prototype from.
- * @throws {TypeError} Will error if either constructor is null, or if
- *     the super constructor lacks a prototype.
- */
-exports.inherits = function(ctor, superCtor) {
+exports._extend = require('internal/extend');
+exports.inherits = require('internal/inherits');
 
-  if (ctor === undefined || ctor === null)
-    throw new TypeError('The constructor to `inherits` must not be ' +
-                        'null or undefined.');
-
-  if (superCtor === undefined || superCtor === null)
-    throw new TypeError('The super constructor to `inherits` must not ' +
-                        'be null or undefined.');
-
-  if (superCtor.prototype === undefined)
-    throw new TypeError('The super constructor to `inherits` must ' +
-                        'have a prototype.');
-
-  ctor.super_ = superCtor;
-  ctor.prototype = Object.create(superCtor.prototype, {
-    constructor: {
-      value: ctor,
-      enumerable: false,
-      writable: true,
-      configurable: true
-    }
-  });
-};
-
-exports._extend = function(origin, add) {
-  // Don't do anything if add isn't an object
-  if (add === null || typeof add !== 'object') return origin;
-
-  var keys = Object.keys(add);
-  var i = keys.length;
-  while (i--) {
-    origin[keys[i]] = add[keys[i]];
-  }
-  return origin;
-};
 
 function hasOwnProperty(obj, prop) {
   return Object.prototype.hasOwnProperty.call(obj, prop);
@@ -829,38 +718,6 @@ exports.pump = internalUtil.deprecate(function(readStream, writeStream, cb) {
 }, 'util.pump is deprecated. Use readableStream.pipe instead.');
 
 
-exports._errnoException = function(err, syscall, original) {
-  var errname = uv.errname(err);
-  var message = syscall + ' ' + errname;
-  if (original)
-    message += ' ' + original;
-  var e = new Error(message);
-  e.code = errname;
-  e.errno = errname;
-  e.syscall = syscall;
-  return e;
-};
-
-
-exports._exceptionWithHostPort = function(err,
-                                          syscall,
-                                          address,
-                                          port,
-                                          additional) {
-  var details;
-  if (port && port > 0) {
-    details = address + ':' + port;
-  } else {
-    details = address;
-  }
-
-  if (additional) {
-    details += ' - Local (' + additional + ')';
-  }
-  var ex = exports._errnoException(err, syscall, details);
-  ex.address = address;
-  if (port) {
-    ex.port = port;
-  }
-  return ex;
-};
+// XXX(bnoordhuis) I think these can just be removed.
+exports._errnoException = require('internal/errno_exception');
+exports._exceptionWithHostPort = require('internal/exception_with_host_port');
