+++ node.gyp
@@ -68,10 +68,18 @@
       'lib/v8.js',
       'lib/vm.js',
       'lib/zlib.js',
+      'lib/internal/assert.js',
       'lib/internal/child_process.js',
+      'lib/internal/debuglog.js',
+      'lib/internal/errno_exception.js',
+      'lib/internal/exception_with_host_port.js',
+      'lib/internal/extend.js',
       'lib/internal/freelist.js',
-      'lib/internal/socket_list.js',
+      'lib/internal/format.js',
+      'lib/internal/inherits.js',
       'lib/internal/repl.js',
+      'lib/internal/socket_list.js',
+      'lib/internal/stream.js',
       'lib/internal/util.js',
     ],
   },
